Move `BackgroundServiceStarter.cs` to WalletWasabi project
